+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6.24 - 2021-10-19
+- Add retryable in case response data invalid json format
+
+## 0.6.22 - 2021-08-17
+- Support Program Members
+
 ## 0.6.21 - 2021-07-15
 - Upgrade `embulk-*` to `v0.10.29`.
 - Apply new lib `embulk-util-*`.